+++ test/account_manager_test.dart
@@ -1,6 +1,6 @@
 import 'package:flutter_test/flutter_test.dart';
-import 'package:account_manager/account_manager.dart';
-import 'package:account_manager/src/generated/account_manager_api.g.dart';
+import 'package:flutter_account_manager/account_manager.dart';
+import 'package:flutter_account_manager/src/generated/account_manager_api.g.dart';
 import 'package:mockito/annotations.dart';
 import 'package:mockito/mockito.dart';